feat: :sparkles: unsubscribe from the auth user snapshot
Now the app receives updates from the auth user in real time, due is not unsubscribing this particular snapshot on route leave. Notice that This also fixes the strange bug that when the user logged in normally, you could not fetch the user information from the database, because it automatically unsubscribed the user (maybe due the route guard).